requirements.txt
pandas
dash
dash-bootstrap-components
sqlalchemy
requests
python-dotenv
openpyxl
fastapi
uvicorn
sqlalchemy
psycopg2-binary
